requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
Werkzeug==0.9.6
itsdangerous==0.24
oauthlib==0.7.2
python-linkedin==4.1
requests==2.5.1
requests-oauthlib==0.4.2
simplejson==3.6.5
wsgiref==0.1.2
